Factored out dictionary of builder methods
